+++ test/jest/__mock__/intl.mock.js
@@ -5,6 +5,30 @@ jest.mock('react-intl', () => {
     formatMessage: ({ id }) => id,
     formatNumber: (value) => value,
     formatDisplayName: (value) => value,
+    formatDate: (value) => value,
+    formatDateToParts: jest.fn(() => ([
+      {
+        'type': 'month',
+        'value': '7'
+      },
+      {
+        'type': 'literal',
+        'value': '/'
+      },
+      {
+        'type': 'day',
+        'value': '31'
+      },
+      {
+        'type': 'literal',
+        'value': '/'
+      },
+      {
+        'type': 'year',
+        'value': '2024'
+      }
+    ]
+    )),
   };
 
   return {